Extract useFocusTrap hook shared by ContentModal and settings modal
ContentModal.tsx and the settings modal in App.tsx contained near-identical
useEffect blocks: Escape key → onClose, Tab/Shift+Tab cycling via the same
querySelectorAll selector, and keydown listener cleanup. Extract into
useFocusTrap(active, modalRef, onClose) in client/src/hooks/useFocusTrap.ts
and replace both duplicates with a single hook call.
